Redirect logged-in users from /day-pass to the portal
Splits /day-pass into a server-component shell that checks the session
and a client DayPassMarketing component holding the existing pricing UI.
